Implement grouping in new details view and column view
